Cleaned up includes and precompiled headers
Former-commit-id: 9504b9a758db96cd01ee6cf33c48a334a39251e1 [formerly 9ffbe1815ffb6de57631cb76aaa6f9a5a2e5196e]
Former-commit-id: ddceccc5c25d2b2702b6d6273c7327b5d12de818

+++ fancy_core/DynamicDescriptorHeapDX12.cpp
@@ -1,4 +1,4 @@
-#include "FancyCorePrerequisites.h"
+#include "fancy_core_precompile.h"
 #include "DynamicDescriptorHeapDX12.h"
 
 #include "DescriptorDX12.h"
@@ -7,7 +7,7 @@
 
 namespace Fancy {
 //---------------------------------------------------------------------------//
-  DynamicDescriptorHeapDX12::DynamicDescriptorHeapDX12(D3D12_DESCRIPTOR_HEAP_TYPE aType, uint64 aNumDescriptors)
+  DynamicDescriptorHeapDX12::DynamicDescriptorHeapDX12(D3D12_DESCRIPTOR_HEAP_TYPE aType, uint aNumDescriptors)
     : myNextFreeHandleIndex(0u)
     , myHandleIncrementSize(0u)
   {
